requirements.txt
django==1.8
psycopg2
Pillow
django-lockout
